+++ docs/.vitepress/config.ts
@@ -51,6 +51,7 @@ export default defineConfig({
                         text: 'Heterogeneous Collections',
                         link: '/collections',
                     },
+                    { text: 'Localization', link: '/localization' },
                     { text: 'OpenAPI Definitions', link: '/openapi' },
                 ],
             },