+++ src/Endpoint/Concerns/FindsResources.php
@@ -25,7 +25,13 @@ trait FindsResources
         }
 
         if (!($model = $collection->find($id, $context))) {
-            throw new ResourceNotFoundException($collection->name(), $id);
+            throw new ResourceNotFoundException(
+                $collection->name(),
+                $id,
+                $context->translate('resource.not_found', [
+                    'identifier' => $collection->name() . '.' . $id,
+                ]),
+            );
         }
 
         return $model;