+++ src/Translation/TranslatorInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Tobyz\JsonApiServer\Translation;
+
+interface TranslatorInterface
+{
+    public function translate(string $key, array $replacements = []): string;
+}